add missing localization keys, move strings to /imports/locales
