+++ imports/locales/strings.js
@@ -3,11 +3,12 @@ export const strings = {
     'app.welcome': 'Did you know that Germany spends {amount} billion Euros every year on public broadcasting?  Financed by monthly contributions of the German public.',
     'app.purpose': 'However, nobody has a say in the programme design. We want to change that: with our tool, you can make your voice heard and propose on which shows your money should be spent.',
     'app.filter.select-filter': 'There are so many programs available. In order to narrow down choices, click on the filter options below.',
+    'app.filter.select-filter.title': 'Filters',
     'app.filter.select-filter.label.category': 'by category',
     'app.filter.select-filter.result-set-size': '{programCounter, plural, =0 {no program matches} one {# program matches} other {# programs match}} the criteria',
     'app.filter.continue-to-decide': 'See relevant programs',
     'app.decide.continue-to-assign': 'Assign money',
-    'app.decide.remaining-programs': '{programCounter, plural, =0 {Finished} one {# one program remains} other {# programs remain}}',
+    'app.decide.remaining-programs': '{programCounter, plural, =0 {Finished} one {# program remains} other {# programs remain}}',
     'app.decide.help-message.intention': 'Now, you can decide whether you want to give money to a program.',
     'app.decide.help-message.todo': 'Simply click on {yes} or {no}. If you are not sure, choose {no}.',
     'app.decide.answer.yes': 'Yes',
@@ -18,28 +19,31 @@ export const strings = {
     'app.assign.restart': 'Find even more programs',
     'app.assign.invoice-table.programs': '{selectionCounter, plural, one {# program} other {# programs}}',
     'app.assign.invoice-table.amount': 'Amount',
+    'app.footer.feedback': 'Want to give us Feedback?',
 
 
   },
 
   'de-DE':{
     'app.welcome': 'Wusstest Du, dass Deutschland jedes Jahr {amount} Milliarden Euro für den öffentlichen Rundfunk ausgibt? Finanziert wird das unter Anderem mit Deinem monatlichen Rundfunkbeitrag.',
-    'app.purpose': 'Aber keiner darf mitbestimmen. Das wollen wir ändern: Mit unserem Tool kannst Du Deiner Stimme Gehör verschaffen und Du selbst kannst darüber entscheiden, wohin Deine Gebühren fließen sollen.',
-    'app.filter.select-filter': 'Es gibt so viele unterschiedliche Programme. Um das Angebot etwas einzuschränken, klick unten auf die Filter-Optionen.',
+    'app.purpose': 'Schade, dass keiner mitbestimmen darf. Das wollen wir ändern: Mit unserem Tool kannst Du Deiner Stimme Gehör verschaffen und Du selbst kannst darüber entscheiden, wohin Deine Gebühren fließen sollen.',
+    'app.filter.select-filter': 'Es gibt so viele unterschiedliche Programme. Um das Angebot etwas einzuschränken, klick einfach unten auf die Filter-Optionen.',
+    'app.filter.select-filter.title': 'Filter',
     'app.filter.select-filter.label.category': 'nach Rubrik',
     'app.filter.select-filter.result-set-size': '{programCounter, plural, =0 {kein Program erfüllt} one {# Programm erfüllt} other {# Programme erfüllen}} die Kriterien',
     'app.filter.continue-to-decide': 'Relevante Programme anschauen',
     'app.decide.continue-to-assign': 'Geld zuweisen',
     'app.decide.remaining-programs': '{programCounter, plural, =0 {Fertig} one {# Programm übrig} other {# Programme übrig }}',
-    'app.decide.help-message.intention': 'Jetzt kannst Du entscheiden, ob Du ein Programm bezahlen möchtest oder nicht.',
+    'app.decide.help-message.intention': 'Hier kannst Du entscheiden, ob Du ein Programm bezahlen möchtest oder nicht.',
     'app.decide.help-message.todo': 'Klick einfach auf {yes} oder {no}. Wenn Du nicht sicher bist, klick auf {no}.',
     'app.decide.answer.yes': 'Ja',
     'app.decide.answer.no': 'Nein',
     'app.assign.title': 'Zusammenfassung',
-    'app.assign.help-message.intention': 'Das war\'s! Bald kannst Du auf dieser Seite angeben wieviel für jedes Programm ausgegeben werden soll. Es wird möglich sein, Beträge auf Gruppen von Programmen zu verteilen oder einzelnen Programmen genau zuzuweisen.',
-    'app.assign.help-message.todo': 'Für\'s erste verteilen wir dein Budget von {budget} pro Monat gleich auf alle ausgewählte Programme.',
+    'app.assign.help-message.intention': 'Das war es schon! Bald kannst Du auf dieser Seite angeben wieviel Du für jedes Programm ausgegeben willst. Es wird möglich sein Beträge auf Gruppen von Programmen zu verteilen oder einzelnen Programmen genau zuzuweisen.',
+    'app.assign.help-message.todo': 'Deswegen verteilen wir erst einmal dein Budget von {budget} pro Monat gleichmäßig auf alle ausgewählten Programme.',
     'app.assign.restart': 'Noch mehr Programme finden',
     'app.assign.invoice-table.programs': '{selectionCounter, plural, one {# Programm} other {# Programme}}',
     'app.assign.invoice-table.amount': 'Betrag',
+    'app.footer.feedback': 'Möchtest Du uns Feedback geben?',
   }
 };
